+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixed not working with the latest version of `resolve`.
+- Fixed remapping throwing an exception if a main module contained an exported variable.
 
 # [4.2.0] - 2020-01-28